% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMEval.R
\name{aStarDM}
\alias{aStarDM}
\title{aStarDM}
\usage{
aStarDM(roads, car, packages)
}
\arguments{
\item{roads}{See runDeliveryMan function in DeliveryMan project for details}

\item{car}{See runDeliveryMan function in DeliveryMan project for details}

\item{packages}{See runDeliveryMan function in DeliveryMan project for details}
}
\value{
packages See runDeliveryMan function in DeliveryMan project for details
}
\description{
The par function for the delivery man project.
}
